services/todo-api/requirements.txt
fastapi
uvicorn[standard]
protobuf
grpcio
grpcio-tools